work saved
work saved

+++ Data/Scripts/ServerCleaner/Configuration.cs
@@ -7,7 +7,7 @@ namespace ServerCleaner
 
         public bool DeletePirates_Enabled = false;
         public int DeletePirates_Interval = 0;
-        public int DeletePirates_PlayerId = 0;
+        public long DeletePirates_PlayerId = 0;
 
         public bool StopAllShips_Enabled = true;
         public int StopAllShips_Interval = 40 * 60 * 1000;
